Refactor value storage
Pass values to the kv store as a byte slice
rather than an owned vec.  The version table
copies these bytes to an internal byte vector.

This ensures that values are tightly packed
in memory, improving cache locality.

It will also enable us to generalize the values
from bytes to custom types in the future.